Revert "Nav-components restructured and made nav bar responsive"
