+++ src/App.jsx
@@ -9,9 +9,8 @@ import Login from './components/login-signup/Login';
 import Signup from './components/login-signup/SignUp';
 import PlaceOrder from './components/order/PlaceOrder';
 // import Order from './pages/Order';
-import './index.css';
-import Navbar from './components/nav-bar/Navbar';
-import Footer from './components/footer/Footer';
+import Navbar from './components/footer-navbar/Navbar';
+import Footer from './components/footer-navbar/Footer';
 import NotFoundPage from './components/home/NotFoundPage';
 import UserProfile from './components/userprofile/UserProfile';
 import ProductPage from './components/products/ProductPage';
@@ -54,7 +53,6 @@ const App = () => {
   return (
     <div className="px-4 sm:px-[5vw] md:px-[7vw] lg:px-[9vw]">
       <Navbar />
-      <main className="main-content">
       <Routes>
         {/* Redirect seller from '/' to '/seller' */}
         <Route
@@ -91,7 +89,7 @@ const App = () => {
 
         <Route path="*" element={<NotFoundPage />} />
       </Routes>
-      </main>        
+
       <Footer />
       <Chatbot />
     </div>
